refactor(ipc,boot): IPC 分域 5 模块 + boot 编排，main.js 定型 119 行（Task 4 + 5.4）
- lib/ipc/{sender,app,recovery,plugin,onboard,session,index}.ts：registerChromeIpc（565 行）按域拆分，channel 名与行为逐一保留；来源校验集中 fromMainWindow/fromWizardWindow
- lib/boot.ts：boot 全链（日志 init→IPC/托盘→新老判定→看门狗/崩溃自回退→渲染自恢复→迁移→向导→插件同步→koffi 预检→市场排队→完整性校验→守护启动→健康确认/定时器）+ handleBootFailure 多级失败链 + fatal + verifyBundledModules
- main.js 重写为装配入口（组合根）：bridge 12 项注入 + 单实例锁/退出清理生命周期，4424→119 行（原目标 <400）
- recovery-integration 断言按新架构指向 lib/ipc/recovery、lib/boot 等（markCleanExit 计 quit 全路径）
- 环境修复：electron dist 缺失导致 require('electron') 自动派生 install.js 挂起测试 —— 以 ELECTRON_MIRROR=npmmirror 重装 electron 二进制解决
- 验证：npm test 450/450 全绿（含 pretest 构建）、typecheck 零错、check-syntax 通过
