+++ dsh-desktop/test/recovery-integration.test.mjs
@@ -1,7 +1,10 @@
 // TDD wiring tests: the recovery/watchdog modules must actually be wired
-// into the desktop shell. main.js is an Electron entry (untestable under
-// node:test directly), so we pin the wiring points at the source level —
-// each assertion corresponds to a required integration point.
+// into the desktop shell. Electron 入口不可在 node:test 下直接执行，故在
+// 源码层面钉住接线点 —— 每条断言对应一个必需的集成位。
+//
+// Task 4/5.4：装配入口 main.js 只剩 bridge 注入 + 生命周期；以下符号分别
+// 位于 lib/run-state.ts、lib/watchdog-boot.ts、lib/window.ts、lib/ipc/
+// recovery.ts、lib/boot.ts，入口/装配断言改为对应模块的 require 调用。
 
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
@@ -12,49 +15,49 @@ import { join } from 'node:path';
 const ROOT = join(fileURLToPath(import.meta.url), '..', '..');
 const mainSrc = readFileSync(join(ROOT, 'main.js'), 'utf8');
 const preloadSrc = readFileSync(join(ROOT, 'preload.js'), 'utf8');
-// Task 3：渲染自恢复装配（initRendererRecovery/wireWindowRecovery/心跳轮询）
-// 迁 lib/window.ts。
 const windowSrc = readFileSync(join(ROOT, 'lib', 'window.ts'), 'utf8');
+const runStateSrc = readFileSync(join(ROOT, 'lib', 'run-state.ts'), 'utf8');
+const watchdogSrc = readFileSync(join(ROOT, 'lib', 'watchdog-boot.ts'), 'utf8');
+const recoveryIpcSrc = readFileSync(join(ROOT, 'lib', 'ipc', 'recovery.ts'), 'utf8');
+const bootSrc = readFileSync(join(ROOT, 'lib', 'boot.ts'), 'utf8');
 
-test('main.js requires the renderer-recovery module', () => {
+test('renderer-recovery is imported and the state machine attaches the main window', () => {
   // Task 3：lib/window.ts 以 ESM import 引 renderer-recovery.js（编译为 require）。
   assert.ok(/from '\.\.\/renderer-recovery\.js'/.test(windowSrc), "lib/window.ts must import '../renderer-recovery.js'");
-});
-
-test('main.js builds the recovery state machine and attaches the main window', () => {
   assert.ok(/export function initRendererRecovery\(\)/.test(windowSrc), 'initRendererRecovery() missing');
-  // Task 1.1：顶层状态迁 lib/state.ts 单例后，引用统一为 state.recovery / state.mainWindow。
   assert.ok(/state\.recovery\.attach\(state\.mainWindow,\s*'main'\)/.test(windowSrc), 'main window attach missing');
 });
 
-test('main.js runs the watchdog lifecycle: run-state write, spawn, clean-exit mark', () => {
-  // Task 2：run-state/writeRunState/markCleanExit 迁 lib/run-state.ts，
-  // startWatchdog 迁 lib/watchdog-boot.ts；main.js 经 require 接线并在 boot 链调用。
-  const runStateSrc = readFileSync(join(ROOT, 'lib', 'run-state.ts'), 'utf8');
-  const watchdogSrc = readFileSync(join(ROOT, 'lib', 'watchdog-boot.ts'), 'utf8');
+test('watchdog lifecycle: run-state write, spawn, clean-exit mark', () => {
   assert.ok(/export function writeRunState\(/.test(runStateSrc), 'writeRunState() missing');
   assert.ok(/export function markCleanExit\(/.test(runStateSrc), 'markCleanExit() missing');
   assert.ok(/export function startWatchdog\(\)/.test(watchdogSrc), 'startWatchdog() missing');
-  assert.ok(/require\('\.\/lib\/watchdog-boot\.js'\)/.test(mainSrc), 'watchdog-boot wiring missing');
-  assert.ok(/startWatchdog\(\);/.test(mainSrc), 'startWatchdog() is never called');
+  // boot 链负责调用（Task 5.4 起 watchdog 生命周期在 lib/boot.ts）。
+  assert.ok(/startWatchdog\(\);/.test(bootSrc), 'startWatchdog() is never called');
 });
 
-test('main.js registers the heartbeat IPC and polls heartbeats', () => {
-  assert.ok(mainSrc.includes("'dsh:renderer-heartbeat'"), 'heartbeat IPC channel missing');
-  // Task 3：心跳轮询迁 lib/window.ts 的 startHeartbeatLoop。
+test('heartbeat IPC is registered and heartbeats are polled', () => {
+  assert.ok(recoveryIpcSrc.includes("'dsh:renderer-heartbeat'"), 'heartbeat IPC channel missing');
   assert.ok(/checkHeartbeats\(\)/.test(windowSrc), 'checkHeartbeats() loop missing');
 });
 
-test('main.js serves the local recovery page IPC endpoints', () => {
+test('the local recovery page IPC endpoints are served', () => {
   for (const ch of ['chrome:recovery-state', 'chrome:recovery-reload', 'chrome:recovery-restart', 'chrome:export-logs']) {
-    assert.ok(mainSrc.includes(`'${ch}'`), `IPC handler ${ch} missing`);
+    assert.ok(recoveryIpcSrc.includes(`'${ch}'`), `IPC handler ${ch} missing`);
   }
   assert.ok(existsSync(join(ROOT, 'assets', 'recovery.html')), 'assets/recovery.html missing');
 });
 
 test('every quit path marks a clean exit for the watchdog', () => {
-  const marks = mainSrc.match(/markCleanExit\(\)/g) || [];
-  assert.ok(marks.length >= 3, `expected markCleanExit() on before-quit + restart + app.exit paths, found ${marks.length}`);
+  // 退出路径分散在 main.js（before-quit）、lib/boot.ts（fatal/校验失败）、
+  // lib/update-flow.ts（agent/客户端更新重启）—— 全部源码合计 ≥3。
+  const sources = [mainSrc, bootSrc];
+  const updateFlowSrc = readFileSync(join(ROOT, 'lib', 'update-flow.ts'), 'utf8');
+  sources.push(updateFlowSrc);
+  const marks = sources.reduce(
+    (n, src) => n + (src.match(/markCleanExit\(\)/g) || []).length, 0,
+  );
+  assert.ok(marks >= 3, `expected markCleanExit() on before-quit + restart + app.exit paths, found ${marks}`);
 });
 
 test('preload sends renderer heartbeats and exposes the recovery bridge', () => {
@@ -62,4 +65,9 @@ test('preload sends renderer heartbeats and exposes the recovery bridge', () => 
   for (const ch of ['chrome:recovery-state', 'chrome:recovery-reload', 'chrome:recovery-restart', 'chrome:export-logs']) {
     assert.ok(preloadSrc.includes(`'${ch}'`), `preload bridge for ${ch} missing`);
   }
+});
+
+test('main.js wires the boot module and registers IPC through lib/ipc', () => {
+  assert.ok(/require\('\.\/lib\/boot\.js'\)/.test(mainSrc), 'boot wiring missing');
+  assert.ok(/registerIpc\(\);/.test(bootSrc), 'registerIpc() is never called');
 });
